ShooterAgent remade
Due to some errors in the agent's behaviour, its script has been totatlly remade from the beggining, as well as the area's.

The agent has been trained to have a basic behaviour. It only aims to the enemy.

The shooterArea is now prepared to deal with multiple agents and enemies at the same time.